requirements.txt
Flask==3.0.2
Werkzeug==3.0.1
pytz==2025.1
flask-wtf==1.2.1
wtforms==3.1.1
Jinja2==3.1.2
click==8.1.7
blinker==1.7.0
itsdangerous==2.1.2
markupsafe==2.1.3
python-dotenv==1.0.1
pandas==2.2.1
python-docx==1.1.0
numpy==1.26.4
